requirements.txt
whatsapp-python==4.3.0
Flask==3.1.0
python-dotenv==1.0.1
gspread==6.1.4
openai==1.58.1